Add nexus staging configuration in the profile section.
For some weird reason, latest release to maven central failed, with the
staging repository not released after being closed. I'm not really sure
why, but maybe configuration isn't propagated as I think it should be.
So let's try adding the nexus-staging configuration to the profile and
be explicit about it.

Change-Id: I30da7185b8cb30606b70ad6ac349b3d548fd5df7