Removed "maven-central" profile.
This profile was active by default and the plan was to de-activate it
for releases to WMF Archiva repository. This does not seem to be needed:
we have done a few releases to WMF Archiva without disabling the
"maven-central" profile and it works fine. Removing this profile
reduces complexity without changing the functionnality.

Change-Id: Ic9427f3f19ffd17b1ee42114e43d9091b626684c