Release to central done via profile.
The previous approach was to configure both deploy plugin and sonatype
nexus-staging-maven-plugin, and have an option to disable nexus. It
seems more straight forward to only activate the nexus plugin when
needed.

So the requirements to push to central are now all in the
"deploy-central" plugin, and the "deploy-archiva" plugin has been
removed. Deployments to central need to use "-P deploy-central",
deployments to archiva don't need any profile.

Change-Id: I72814ece7b9b619f0677aa62d20321d51b106fb3