Add sonar configuration to analyze checkstyle and spotbugs reports.
In most projects, we fail the build directly on checkstyle and
spotbugs violations, so sonar analysis is not needed. But in some
cases (legacy projects that have too many violations), we don't
want to fail the build, but we want to report violations so that
they can be addressed over time.

Change-Id: I4733b4db5e084a9d4235ca898efc503ffa2d005b