Update urls to match gerrit updates
The top-level url was missing `wikimedia` in the name, leading to
an error page instead of the project page. Update the url to match
reality.

Gerrit has changed since this was written and now uses full urls
instead of placing some of the url after the hash (#). Update to
use the modern url.

Change-Id: Iba78ab214c4a81c3060b79b202f6685b40c48a9c